refactor: center dto 패키지 분리 및 api 패키지 package-info 파일 작성
