+++ daycan-back/api/src/main/java/com/daycan/api/controller/member/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * member 계정을 위한 rest API 관련 controller 패키지입니다.
+ */
+package com.daycan.api.controller.member;
