+++ daycan-back/api/src/main/java/com/daycan/api/dto/center/request/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Center API 의 요청 DTO가 포함된 패키지입니다.
+ */
+
+package com.daycan.api.dto.center.request;
