--- daycan-back/api/src/main/java/com/daycan/api/dto/center/response/DocumentCountResponse.java
@@ -1,8 +1,0 @@
-package com.daycan.api.dto.center.response;
-
-public record DocumentCountResponse(
-    CareReportCountResponse careReport,
-    CareSheetCountResponse careSheet
-) {
-
-}
