+++ daycan-back/api/src/main/java/com/daycan/api/dto/center/response/centermanage/AttendanceResultResponse.java
@@ -1,4 +1,4 @@
-package com.daycan.api.dto.center.response;
+package com.daycan.api.dto.center.response.centermanage;
 
 import io.swagger.v3.oas.annotations.media.Schema;
