+++ daycan-back/api/src/main/java/com/daycan/api/dto/center/response/centermanage/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * 센터 관리 업무 (종사자, 수급자 관리) 등을 위한 응답 DTO 패키지입니다.
+ */
+
+package com.daycan.api.dto.center.response.centermanage;
