+++ daycan-back/api/src/main/java/com/daycan/api/dto/center/response/document/DocumentCountResponse.java
@@ -1,0 +1,11 @@
+package com.daycan.api.dto.center.response.document;
+
+import com.daycan.api.dto.center.response.report.CareReportCountResponse;
+import com.daycan.api.dto.center.response.sheet.CareSheetCountResponse;
+
+public record DocumentCountResponse(
+    CareReportCountResponse careReport,
+    CareSheetCountResponse careSheet
+) {
+
+}
