+++ daycan-back/api/src/main/java/com/daycan/api/dto/center/response/document/DocumentStatusResponse.java
@@ -1,4 +1,4 @@
-package com.daycan.api.dto.center.response;
+package com.daycan.api.dto.center.response.document;
 
 import com.daycan.domain.entry.document.report.ReportStatus;
 import com.daycan.domain.entry.document.sheet.SheetStatus;
