+++ daycan-back/api/src/main/java/com/daycan/api/dto/center/response/document/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 문서 (기록지/리포트)통합 관리를 위한 DTO 패키지입니다.
+ */
+package com.daycan.api.dto.center.response.document;
