+++ daycan-back/api/src/main/java/com/daycan/api/dto/center/response/image/ProfileUploadResponse.java
@@ -1,4 +1,4 @@
-package com.daycan.api.dto.center.response;
+package com.daycan.api.dto.center.response.image;
 
 public record ProfileUploadResponse(String originKey, String thumbKey) {}
