+++ daycan-back/api/src/main/java/com/daycan/api/dto/center/response/image/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 이미지, 사진, OCR 관련 응답 DTO 패키지입니다.
+ */
+package com.daycan.api.dto.center.response.image;
