+++ daycan-back/api/src/main/java/com/daycan/api/dto/center/response/report/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * center API 의 리포트 관련 응답 DTO가 포함된 패키지입니다.
+ */
+package com.daycan.api.dto.center.response.report;
