+++ daycan-back/api/src/main/java/com/daycan/api/dto/center/response/sheet/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 기록지 관련 응답 DTO 패키지
+ */
+package com.daycan.api.dto.center.response.sheet;
