+++ daycan-back/api/src/main/java/com/daycan/api/dto/common/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * 이 패키지는 공통적으로 사용되는 DTO(Data Transfer Object) 클래스들을 포함합니다.
+ * Center API와 Member API에서 공통적으로 사용되는 요청 및 응답 DTO가 포함되어 있습니다.
+ */
+
+package com.daycan.api.dto.common;
