+++ daycan-back/api/src/main/java/com/daycan/api/dto/member/response/statistics/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 이 패키지는 Member 의 통계 관련 응답 DTO 클래스를 포함합니다.
+ */
+package com.daycan.api.dto.member.response.statistics;
