+++ daycan-back/api/src/main/java/com/daycan/api/dto/validation/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * 이 패키지는 DTO 검증을 위한 어노테이션을 포함합니다.
+ */
+
+package com.daycan.api.dto.validation;
