+++ daycan-back/api/src/main/java/com/daycan/external/S3Service.java
@@ -1,7 +1,7 @@
 package com.daycan.external;
 
 import com.daycan.common.response.status.error.CommonErrorStatus;
-import com.daycan.api.dto.center.response.PresignResponse;
+import com.daycan.api.dto.center.response.image.PresignResponse;
 import com.daycan.common.exceptions.ApplicationException;
 import lombok.RequiredArgsConstructor;
