+++ daycan-back/api/src/main/java/com/daycan/service/center/StaffService.java
@@ -7,7 +7,7 @@ import com.daycan.domain.entity.Staff;
 import com.daycan.domain.enums.Gender;
 import com.daycan.domain.enums.StaffRole;
 import com.daycan.api.dto.center.request.AdminStaffRequest;
-import com.daycan.api.dto.center.response.AdminStaffResponse;
+import com.daycan.api.dto.center.response.centermanage.AdminStaffResponse;
 import com.daycan.repository.jpa.CenterRepository;
 import com.daycan.repository.jpa.StaffRepository;
 import java.util.List;
