+++ daycan-back/api/src/main/java/com/daycan/service/document/DocumentFacade.java
@@ -1,12 +1,11 @@
 package com.daycan.service.document;
 
 import com.daycan.api.dto.center.request.AttendanceAction;
-import com.daycan.api.dto.center.request.AttendanceMarkRequest;
 import com.daycan.api.dto.center.request.CareSheetRequest;
-import com.daycan.api.dto.center.response.AttendanceResultResponse;
-import com.daycan.api.dto.center.response.CareSheetMetaResponse;
-import com.daycan.api.dto.center.response.CareSheetResponse;
-import com.daycan.api.dto.center.response.DocumentStatusResponse;
+import com.daycan.api.dto.center.response.centermanage.AttendanceResultResponse;
+import com.daycan.api.dto.center.response.sheet.CareSheetMetaResponse;
+import com.daycan.api.dto.center.response.sheet.CareSheetResponse;
+import com.daycan.api.dto.center.response.document.DocumentStatusResponse;
 import com.daycan.common.exceptions.DocumentNonCreatedException;
 import com.daycan.common.response.PageResponse;
 import com.daycan.domain.entry.document.sheet.SheetStatus;
@@ -21,7 +20,6 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.EnumSet;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
